requirements.txt
# Core LangGraph dependencies
langgraph>=0.2.6
langgraph-cli[inmem]>=0.1.0
langgraph-sdk>=0.1.0
langgraph-api>=0.1.0
langgraph-checkpoint-postgres>=0.1.0

# Python environment
python-dotenv>=1.0.1

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
starlette>=0.37.0

# Authentication
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.7

# HTTP client
httpx>=0.27.0

# LLM providers & LangChain integrations
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langchain-google-genai>=1.0.0
langchain-text-splitters>=0.2.0
langchain-anthropic>=0.1.0
langchain-groq>=0.1.0
openai>=1.12.0
groq>=0.5.0

# Vector database
qdrant-client>=1.9.0

# Database - PostgreSQL (complete stack)
sqlalchemy>=2.0.30
# Use both psycopg2-binary and psycopg to ensure compatibility
psycopg2-binary>=2.9.0
psycopg[binary,pool]>=3.1.18
alembic>=1.13.1
asyncpg>=0.29.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# Date/time utilities
pytz>=2023.3

# Google services (for GenAI & TTS)
#google-generativeai>=0.5.0
google-api-python-client==2.125.0
google-cloud-texttospeech>=2.16.0

# Search APIs
#tavily-python>=0.3.3
## langchain-tavily removed due to conflict with langchain_tavily
# SerpAPI client
google-search-results>=2.4.0

# Utilities
pydantic>=2.7.0
pydantic-settings>=2.2.0
typing-extensions>=4.10.0

langchain_tavily
langmem
google-generativeai